refactor: 重构component control #1821
