+++ packages/component/src/control/baseControl/buttonControl.ts
@@ -1,5 +1,5 @@
+import { Control, IControlOption } from '@antv/l7-core';
 import { DOM } from '@antv/l7-utils';
-import Control, { IControlOption } from './control';
 
 export { ButtonControl };
