+++ packages/component/src/control/baseControl/index.ts
@@ -1,5 +1,5 @@
+export { Control, PositionType, type IControlOption } from '@antv/l7-core';
 export { ButtonControl, type IButtonControlOption } from './buttonControl';
-export { Control, PositionType, type IControlOption } from './control';
 export { PopperControl, type IPopperControlOption } from './popperControl';
 export {
   SelectControl,
